RF: cmdline - extract code into helper functions, move into helpers
main:setup_parser was hard to digest primarily because of the large
blocks of code combined into a single long spaggetti, without clearly
identifiable purpose.  I have extracted those blocks into "protected"
functions, and moved them into helpers to keep main.py relatively
succinct.  I also moved other used by them functions into helpers.py
as well to avoid circular imports etc.  Corresponding tests were moved
as well
