+++ datalad/cmdline/helpers.py
@@ -12,10 +12,12 @@
 __docformat__ = 'restructuredtext'
 
 import argparse
+import logging
 import os
 import re
 import sys
 import gzip
+import textwrap
 import warnings
 from tempfile import NamedTemporaryFile
 from textwrap import wrap
@@ -23,13 +25,20 @@ from textwrap import wrap
 from ..cmd import WitlessRunner as Runner
 from ..interface.common_opts import eval_defaults
 from ..log import is_interactive
+from ..ui.utils import get_console_width
 from ..utils import (
     ensure_unicode,
     getpwd,
-    unlink,
+    unlink, get_suggestions_msg,
 )
-from ..version import __version__
-from ..dochelpers import exc_str
+from ..version import __version__, __full_version__
+from ..dochelpers import exc_str, exc_str_old
+
+from appdirs import AppDirs
+from os.path import join as opj
+
+dirs = AppDirs("datalad", "datalad.org")
+
 
 from logging import getLogger
 lgr = getLogger('datalad.cmdline')
@@ -125,13 +134,23 @@ class LogLevelAction(argparse.Action):
         LoggerHelper().set_level(level=values)
 
 
-# MIH: Disabled. Non-functional, untested.
-#class PBSAction(argparse.Action):
-#    """Action to schedule actual command execution via PBS (e.g. Condor)"""
-#    def __call__(self, parser, namespace, values, option_string=None):
-#        pbs = values[0]
-#        import pdb; pdb.set_trace()
-#        i = 1
+class ArgumentParserDisableAbbrev(argparse.ArgumentParser):
+    # Don't accept abbreviations for long options. This kludge was originally
+    # added at a time when our minimum required Python version was below 3.5,
+    # preventing us from using allow_abbrev=False. Now our minimum Python
+    # version is high enough, but we still can't use allow_abbrev=False because
+    # it suffers from the problem described in 6b3f2fffe (BF: cmdline: Restore
+    # handling of short options, 2018-07-23).
+    #
+    # Modified from the solution posted at
+    # https://bugs.python.org/issue14910#msg204678
+    def _get_option_tuples(self, option_string):
+        chars = self.prefix_chars
+        if option_string[0] in chars and option_string[1] in chars:
+            # option_string is a long flag. Disable abbreviation.
+            return []
+        return super(ArgumentParserDisableAbbrev, self)._get_option_tuples(
+            option_string)
 
 
 def parser_add_version_opt(parser, mod_name, include_name=False, delay=False):
@@ -321,11 +340,7 @@ queue
         unlink(f.name)
 
 
-# TODO: useful also outside of cmdline, move to support/
-from os import curdir
-
-
-def get_repo_instance(path=curdir, class_=None):
+def get_repo_instance(path=os.curdir, class_=None):
     """Returns an instance of appropriate datalad repository for path.
     Check whether a certain path is inside a known type of repository and
     returns an instance representing it. May also check for a certain type
@@ -388,8 +403,267 @@ def get_repo_instance(path=curdir, class_=None):
     else:
         raise RuntimeError("No datalad repository found in %s" % abspath_)
 
+#
+# Some logic modules extracted from main.py to de-spagetify
+#
 
-from appdirs import AppDirs
-from os.path import join as opj
+def _maybe_get_single_subparser(cmdlineargs, parser, interface_groups,
+                                return_subparsers, completing, help_ignore_extensions):
+    """Performs early analysis of the cmdline
 
-dirs = AppDirs("datalad", "datalad.org")
+    Looks at the first unparsed argument and if a known command, would return_subparsers
+    is False, would return only that one.
+
+    For the analysis to be complete etc, would also load commands from entrypoints etc
+
+    Returns
+    -------
+    None or str
+    """
+    # Before doing anything additional and possibly expensive see may be that
+    # we have got the command already
+    need_single_subparser = False if return_subparsers else None
+    fail_handler = (lambda *a, **kw: True) \
+        if return_subparsers else fail_with_short_help
+    try:
+        parsed_args, unparsed_args = parser._parse_known_args(
+            cmdlineargs[1:], argparse.Namespace())
+        # before anything handle possible datalad --version
+        if not unparsed_args and getattr(parsed_args, 'version', None):
+            parsed_args.version()  # will exit with 0
+        if not (completing or unparsed_args):
+            fail_handler(parser, msg="too few arguments", exit_code=2)
+        lgr.debug("Command line args 1st pass for DataLad %s. Parsed: %s Unparsed: %s",
+                  __full_version__, parsed_args, unparsed_args)
+    except Exception as exc:
+        lgr.debug("Early parsing failed with %s", exc_str_old(exc))
+        need_single_subparser = False
+        unparsed_args = cmdlineargs[1:]  # referenced before assignment otherwise
+    # First unparsed could be either unknown option to top level "datalad"
+    # or a command. Among unknown could be --help/--help-np which would
+    # need to be dealt with
+    unparsed_arg = unparsed_args[0] if unparsed_args else None
+    if need_single_subparser is not None \
+            or unparsed_arg in ('--help', '--help-np', '-h'):
+        need_single_subparser = False
+        if not help_ignore_extensions:
+            add_entrypoints_to_interface_groups(interface_groups)
+    elif not completing and unparsed_arg.startswith('-'):  # unknown option
+        fail_with_short_help(parser,
+                             msg="unrecognized argument %s" % unparsed_arg,
+                             exit_code=2)
+        # if we could get a list of options known to parser,
+        # we could suggest them
+        # known=get_all_options(parser), provided=unparsed_arg)
+    else:  # the command to handle
+        known_commands = get_commands_from_groups(interface_groups)
+        if unparsed_arg not in known_commands:
+            # need to load all the extensions and try again
+            add_entrypoints_to_interface_groups(interface_groups)
+            known_commands = get_commands_from_groups(interface_groups)
+
+        if unparsed_arg not in known_commands:
+            # check if might be coming from known extensions
+            from ..interface import (
+                _known_extension_commands,
+                _deprecated_commands,
+            )
+            extension_commands = {
+                c: e
+                for e, commands in _known_extension_commands.items()
+                for c in commands
+            }
+            hint = None
+            if unparsed_arg in extension_commands:
+                hint = "Command %s is provided by (not installed) extension %s." \
+                       % (unparsed_arg, extension_commands[unparsed_arg])
+            elif unparsed_arg in _deprecated_commands:
+                hint_cmd = _deprecated_commands[unparsed_arg]
+                hint = "Command %r was deprecated" % unparsed_arg
+                hint += (" in favor of %r command." % hint_cmd) if hint_cmd else '.'
+            if not completing:
+                fail_with_short_help(
+                    parser,
+                    hint=hint,
+                    provided=unparsed_arg,
+                    known=list(known_commands.keys()) + list(extension_commands.keys())
+                )
+        if need_single_subparser is None:
+            need_single_subparser = unparsed_arg
+    return need_single_subparser
+
+
+def _maybe_get_interface_subparser(_intfspec, subparsers, cmd_name, formatter_class, group_name,
+                                   grp_short_descriptions):
+    """Given an interface spec, add a subparser to subparsers under cmd_name
+
+    That subparser is also gets added to the grp_short_descriptions
+    """
+    from ..interface.base import (
+        alter_interface_docs_for_cmdline,
+        get_cmd_doc,
+        get_cmd_ex,
+        load_interface,
+    )
+    _intf = load_interface(_intfspec)
+    if _intf is None:  # failed to load
+        return
+    # deal with optional parser args
+    if hasattr(_intf, 'parser_args'):
+        parser_args = _intf.parser_args
+    else:
+        parser_args = dict(formatter_class=formatter_class)
+        # use class description, if no explicit description is available
+        intf_doc = get_cmd_doc(_intf)
+        parser_args['description'] = alter_interface_docs_for_cmdline(
+            intf_doc)
+        if hasattr(_intf, '_examples_'):
+            intf_ex = alter_interface_docs_for_cmdline(get_cmd_ex(_intf))
+            parser_args['description'] += intf_ex
+    subparser = subparsers.add_parser(cmd_name, add_help=False, **parser_args)
+    # our own custom help for all commands
+    parser_add_common_opt(subparser, 'help')
+    # let module configure the parser
+    _intf.setup_parser(subparser)
+    # and we would add custom handler for --version
+    parser_add_version_opt(subparser, _intf.__module__.split('.', 1)[0], include_name=True)
+    # logger for command
+    # configure 'run' function for this command
+    plumbing_args = dict(
+        func=_intf.call_from_parser,
+        logger=logging.getLogger(_intf.__module__),
+        subparser=subparser)
+    if hasattr(_intf, 'result_renderer_cmdline'):
+        plumbing_args['result_renderer'] = _intf.result_renderer_cmdline
+    subparser.set_defaults(**plumbing_args)
+    # store short description for later
+    sdescr = getattr(_intf, 'short_description',
+                     parser_args['description'].split('\n')[0])
+    grp_short_descriptions[group_name].append((cmd_name, sdescr))
+    return subparser
+
+
+def add_entrypoints_to_interface_groups(interface_groups):
+    lgr.debug("Loading entrypoints")
+    from pkg_resources import iter_entry_points  # delay expensive import
+    for ep in iter_entry_points('datalad.extensions'):
+        lgr.debug(
+            'Loading entrypoint %s from datalad.extensions for docs building',
+            ep.name)
+        try:
+            spec = ep.load()
+            if len(spec) < 2 or not spec[1]:
+                lgr.debug(
+                    'Extension does not provide a command suite: %s',
+                    ep.name)
+                continue
+            interface_groups.append((ep.name, spec[0], spec[1]))
+            lgr.debug('Loaded entrypoint %s', ep.name)
+        except Exception as e:
+            lgr.warning('Failed to load entrypoint %s: %s', ep.name, exc_str_old(e))
+            continue
+
+
+def get_commands_from_groups(groups):
+    """Get a dictionary of command: interface_spec"""
+    from ..interface.base import get_cmdline_command_name
+    return {
+        get_cmdline_command_name(_intfspec): _intfspec
+        for _, _, _interfaces in groups
+        for _intfspec in _interfaces
+    }
+
+
+def fail_with_short_help(parser=None,
+                         msg=None,
+                         known=None, provided=None,
+                         hint=None,
+                         exit_code=1,
+                         what="command",
+                         out=None):
+    """Generic helper to fail
+    with short help possibly hinting on what was intended if `known`
+    were provided
+    """
+    out = out or sys.stderr
+    if msg:
+        out.write("error: %s\n" % msg)
+    if not known:
+        if parser:
+            # just to appear in print_usage also consistent with --help output
+            parser.add_argument("command [command-opts]")
+            parser.print_usage(file=out)
+    else:
+        out.write(
+            "datalad: Unknown %s %r.  See 'datalad --help'.\n\n"
+            % (what, provided,))
+        if provided not in known:
+            out.write(get_suggestions_msg(provided, known))
+    if hint:
+        out.write("Hint: %s\n" % hint)
+    raise SystemExit(exit_code)
+
+
+def _fix_datalad_ri(s):
+    """Fixup argument if it was a DataLadRI and had leading / removed
+
+    See gh-2643
+    """
+    if s.startswith('//') and (len(s) == 2 or (len(s) > 2 and s[2] != '/')):
+        lgr.info(
+            "Changing %s back to /%s as it was probably changed by MINGW/MSYS, "
+            "see http://www.mingw.org/wiki/Posix_path_conversion", s, s)
+        return "/" + s
+    return s
+
+
+def get_description_with_cmd_summary(grp_short_descriptions, interface_groups,
+                                     parser_description):
+    from ..interface.base import dedent_docstring
+    from ..interface.base import get_cmd_summaries
+    lgr.debug("Generating detailed description for the parser")
+
+    console_width = get_console_width()
+    cmd_summary = get_cmd_summaries(grp_short_descriptions, interface_groups,
+                                    width=console_width)
+    # we need one last formal section to not have the trailed be
+    # confused with the last command group
+    cmd_summary.append('\n*General information*\n')
+    detailed_description = '%s\n%s\n\n%s' \
+                           % (parser_description,
+                              '\n'.join(cmd_summary),
+                              textwrap.fill(dedent_docstring("""\
+    Detailed usage information for individual commands is
+    available via command-specific --help, i.e.:
+    datalad <command> --help"""),
+                                            console_width - 5,
+                                            initial_indent='',
+                                            subsequent_indent=''))
+    return detailed_description
+
+
+def _parse_overrides_from_cmdline(cmdlineargs):
+    """parse config overrides provided in command line
+
+    Might exit(3) the entire process if value is not assigned"""
+    # this expression is deliberately loose as gitconfig offers
+    # quite some flexibility -- this is just meant to catch stupid
+    # errors: we need a section, a variable, and a value at minimum
+    # otherwise we break our own config parsing helpers
+    # https://github.com/datalad/datalad/issues/3451
+    noassign_expr = re.compile(r'[^\s]+\.[^\s]+=[\S]+')
+    noassign = [
+        o
+        for o in cmdlineargs.cfg_overrides
+        if not noassign_expr.match(o)
+    ]
+    if noassign:
+        lgr.error(
+            "Configuration override without section/variable "
+            "or value assignment (must be 'section.variable=value'): %s",
+            noassign)
+        sys.exit(3)
+    overrides = dict([
+        (o.split('=')[0], '='.join(o.split('=')[1:]))
+        for o in cmdlineargs.cfg_overrides])
+    return overrides
